+++ shared/Cargo.toml
@@ -10,3 +10,6 @@ indoc = "2"
 libherokubuildpack.workspace = true
 libcnb.workspace = true
 java-properties = "2"
+
+[dev-dependencies]
+tempfile = "3"
