+++ shared/src/app.rs
@@ -1,0 +1,2 @@
+pub mod maven;
+pub mod sbt;
