+++ shared/src/lib.rs
@@ -6,6 +6,7 @@
 // This lint is too noisy and enforces a style that reduces readability in many cases.
 #![allow(clippy::module_name_repetitions)]
 
+pub mod app;
 pub mod env;
 pub mod fs;
 pub mod log;
